Fix spinner lists from appearing blank.
Spinner lists and regular lists shared the same code and layout.
The problem is that in the default theme the lists have different
backgrounds, and the spinner background is the same colour as the
list text colour.

Distinguish between the two in the code, and add a new layout
item specifically for spinner list items.