Refactor the "now playing" UI.
With the "Home" and "Playlist" buttons moved to the actionbar
the album artwork can be larger.  This necessitated writing
SquareImageView.java for the landscape layout to work properly.

Followed the guidelines at the Android Design site
http://developer.android.com/design/style/metrics-grids.html 
for padding, target size, text size, and so on.

Removed TransparentPanel.java, it's no longer used.