Start NowPlayingFragment.
All the playing + UI logic moves from NowPlayingActivity to
NowPlayingFragment, with the activity existing just to host
the fragment.

In the future the ...Fragment will figure out whether to
display full screen (for NowPlayingActivity) or just take up
a row at the bottom of some other activity that it's hosted
in.

Doesn't work in landscape mode yet, and a handful of known
issues are marked with XXX in the comments.