First cut at implementing the mini-now-playing fragment.
NowPlayingFragment checks the height of the enclosing layout
at onInflate() time, and if it's not fill_parent then it
selects a different layout (now_playing_fragment_mini.xml) to
inflate, and sets a boolean to indicate that it's using the
mini player.

The mini player only provides album artwork, track, and artist
info, so conditionalise a lot of other code in NowPlayingFragment
based on whether it's the mini player or the full player that's
in use.

I'm not sure this is very clean -- I suspect that refactoring
this in to NowPlayingFragmentMini, and a NowPlayingFragmentFull
derived class would result in better code, but this is a work
in progress.

Update item_list.xml to use the new fragment.

While I'm here, add IDs to the fragments, which seems to fix
the orientation issue mentioned in a comment in
NowPlayingActivity, so uncomment that code as well.