Get closer to platform standards for list views.
Adjust the icon sizes and padding in icon_large_row_layout.xml to
be multiples of four, with icons 48dp in size.

Provide drawable-{l,m,h,xh}dpi versions of all icons.  The mdpi icon
images are 48px x 48px, and the others are scaled in the ratio
3:4:6:8 (l:m:h:xh) per 
http://developer.android.com/guide/practices/screens_support.html.

Remove plugin_item_row_layout.xml, it existed to add a right-facing
caret to menu items with submenus, and this is non-standard for Android,
per http://developer.android.com/design/patterns/pure-android.html.