Fix the race condition SqueezerHomeActivity displaying list items.
Introduce the concept of a "handshake" between the service and server,
which is the period where the service queries the server to determine
its capabilities.  The handshake completes when the server responds to
the "version" query.

Introduce a new callback method, onHandshakeCompleted(). Use this in
SqueezerHomeActivity to create the menu list items after the handshake
completes.

Because two things can now have a IServiceCallback registered at the
same time (at least SqueezerHomeActivity and NowPlayingFragment)
change the callback reference from a single AtomicReference to a
RemoteCallbackList, with knock-on changes to how callbacks registered
in that list are called.