Add support for the Squeezeserver discovery protocol.
This drastically speeds up scanning the network for servers.  As a side
effect, we get the server's name, so we can show that to the user
instead of the IP address.

Adjust the progress bar to use the secondary value to indicate scanning
progress, and the primary value to show how many servers have been
found.