Fix 2 bugs when the server address changes.
Bug 1: Shared preference listeners that are anonymous inner classes are at
risk of being garbage collected, and never being called.  Fix this by
extending the activity to implement OnSharedPreferenceChangeListener.

Bug 2: The service never did anything when the server address changed.  It
stayed connected to the old server.  Force the service to disconnect, which
will trigger a reconnection if the activity that's active when we return
from the PreferenceActivity needs to connect.