Use -keepclasseswithmembers.
Per the lint error:

Using -keepclasseswithmembernames in a proguard config file is not
correct; it can cause some symbols to be renamed which should not be.

Earlier versions of ADT used to create proguard.cfg files with the
wrong format. Instead of -keepclasseswithmembernames use
-keepclasseswithmembers, since the old flags also implies "allow
shrinking" which means symbols only referred to from XML and not Java
(such as possibly CustomViews) can get deleted.

http://http://code.google.com/p/android/issues/detail?id=16384