Start fixing how images are downloaded.
The current code in SqueezerIconUpdater.updateIcon is called from
every getView().  getView() may be called multiple times for each
ImageView, and every time it is the image is fetched again.  So
album artwork might be downloaded two or three times for the same
ImageView.

Fix this by introducing an ImageDownloader class.  This class
maintains an in-memory cache of images that have been downloaded,
which means that not only do we stop downloading the same image
multiple times for a single ImageView, we will re-use the image for
any ImageViews on screen that should be associated with the same
image, for an even greater win.

This class is modeled on the SDK example
XmlAdapters/src/com/example/android/xmladapters/ImageDownloader.html,
but modified so that one or more ImageViews are associated with URLs,
and to use LruCache instead of the home-grown approach in that file.

SqueezerSongView has been updated to use this new method, the other
views will be modified in future commits.

Use an animation when images are displayed -- this helps highlight
any times that images are being re-downloaded when they shouldn't
be, for further work.