Show a disclosure triangle on list items that have context menus.
Import the disclosure triangle from the Android SDK (15), as well
as drawables for the button's various states to match the Holo Dark
theme.

Create a subclass of ImageButton to display the triangle, necessary
to avoid unwanted UI effects when the list is long pressed.

Amend the item views that have context menus to display the button,
and show the context menu when the button is pressed.

Adjust some of the layout resources as necessary, and while I'm here,
adjust them to keep within the Android layout design guidelines.