Break the circular link between SqueezerItemAdapter and SqueezerItemView.
SqueezerItemView had a reference to the adapter that created it, which
was used in two specific subclasses to determine whether or not "move
down" entries in the context menu should appear.

Break this link, and explicitly pass the adapter in to the
setupContextMenu() method where it can be used if necessary.