Removed add-to-playlist and play-next from the context-menu of the current playlist.
The standard options of an item in a list of SqueezerSongs includes the ability to add the song to the playlist.
This is probably not what one wants, when viewing the current playlist, since the item is already there. This was previously implemented by not calling through to super.setUpConextMenu.
Since most the the standard functionality is now desired, I removed the items, which don't make sense instead.