Further smoothing of album artwork display.
Implement a 2 level (memory and disk) cache of bitmaps retrieved from
the server to reduce the amount of network activity.

Refactor the scroll handling code to be sensitive to the cache's disk
access, and use an OnTouchListener to track whether the user has their
finger on the screen

Hoist common code out of SqueezerAlbumView and SqueezerSongView, and
put it in SqueezerAlbumArtView.