Make lists with track artwork scroll even more smoothly.
Replace the previous image processing and disk cache code with a slightly
different version from the Google IO Schedule opensource app (with some
bug fixes).

Use an ImageFetcher to fetch images for both the artwork lists and the
"Now Playing" activity.  The ImageFetcher is backed by an in-memory and
on-disk cache.

Pass the ImageFetcher from the list activities, down through the adapter,
and in to the per-item views.

The new code takes account of some quirks in different versions of
Android.  In order to get the correct API constants, increase the
targetSdkVersion to 10, and build against API 16.

Work around the NetworkOnMainThreadExceptions that this can cause by
explicitly allowing networking on the main thread for the time being.