Remove a memory leak when scanning for servers.
Make ScanNetworkTask a static inner class, and cancel the task when
the preference dialog is destroyed (e.g., on screen rotation).

Create a ProgressDialog during the scan and display that, instead of
an embedded ProgressBar in the preference dialog.

Adjust the scanning algorithm slightly, and make it attempt based,
rather than time based.  Drop the number of attempts from ~ 20 to no
more than 5.