Make scrobbling simpler to control.
Instead of asking the user which scrobbling client they'd like to use,
query to see which scrobbling intent broadcast receivers are available,
and use that, favouring ScrobbleDroid over SLS if both are available.

Make the scrobbling preference a simple checkbox, and upgrade from the
user's previous preference if it exists.

Tell the user if they enable scrobbling but no suitable app is installed,
and present them with a list of possible apps to use, linking to Google
Play for ease of installation.