Overflow menu and current playlist changes in the actionbar compatibility code
When running the actionbar-branch on a pre honeycomb device, I have noticed a few issues. This commit solves these.

1) Nothing happens when pressing the menu button. This feels like an error.

2) I can't find the current playlist, because it is hidden in the options menu.

Solution:

The actionbar compatibility code is moving excess actionbar items to the action overflow.
Currently this is a popup menu, which is activated by an overflow icon.
In the native ActionBar, "the overflow icon only appears on phones that have no menu hardware keys. Phones with menu keys display the action overflow when the user presses the key."
Since all pre honeycomb devices have menu keys, and we only use the compatibility code on pre honeycomb devices, it is safe to leave excess actionbar items in the options menu.
This simplifies the code somewhat, while providing additional space on the action bar.

The extra space is used to display an additional icon on the action bar. Additionally the Ifroom option is now supported. The calculation of how many icons to display, is done as described in http://developer.android.com/design/patterns/actionbar.html, except that the restriction of "Action buttons in the main action bar may not occupy more than 50% of the bar's width" isn't checked. The withText option is still not supported.