Disable download of current song when no current song
Both the menu item and the action when the menu item is selected is disabled,
because the player may stop playing while the menu is open. The download is
also disabled if the song is remote.
It is a bit awkward because current song may not be null, even if no song is
playing, but I will fix this in the actionbar branch (which will soon become
master).