Fix out of memory problem with album artwork images
Images for the Squeezer lists are scaled down on download, to avoid problems when the images stored on the server are to large.
Also the image used for no artwork and pending artwork are now dependant on the android device.
This is an intermidiate fix, because in order to find the scale factor, the image needs to be downloaded twice.