Fix list view creation problems
The view generating hierachy and the simple views have become more complex.
This means that it is no longer sufficient to check if the tag is null, to
see if we can reuse the view. We also need to check if the view is the correct
type. We do this by checking the type of the tag, because it leads to
slightly simpler code.