Fix crash in options menu for pre honeycomb devices
Fix crashed which occurs when opening the options menu after a context menu
have been opened.
This error only occurs when running the action bar comptibility code, ie.
on with android versions before 3.0.
We are now using menu resources for context menus, causing inflate to be called
when the context menu are displayed. The compatibility code then had the wrong
data for options menu, and crashed when trying to show/hide menu items, to
reflect whether the controlled squeezbox player was on or off.