Navigation changes
To provide intuitive and consistent navigation the android back stack is
controlled slightly different.

*The home activity: ClearTop|SingleTop
Since the home activity is started first, pressing the home button will display
the home screen and clear the back stack, so pressing the back key will leave
the application.

*Search results activity: singleTop
This means that if the search if an additional search is performed while the
search results activity is active, this instance will be re-used to display the
new results. But if you for example do search -> view albums for a found artist
-> new search, you can press back twice to get to the previous search results.

*Now playing, current playlist and player list: FLAG_ACTIVITY_REORDER_TO_FRONT
Multiple instances of these activites doesn't make sense, so we don't want
that. If the activity is allready running it will be brought to the foreground,
otherwise it will be started. In any case pressing the back key will take you
to the previous activity.

*All other lists: android standard.
Every time a list is seleced, a new instance is created, which holds the data.
So you can for example do: home -> artists -> albums by an artist -> artists
for an album, and go back through each activity