Only add resized bitmaps to the memory cache, keep originals on disk.
Since the disk cache is shared between the activities that show
thumbnails and the "Now Playing" activity this ensures that the "Now
Playing" activity doesn't show resized thumbnails of the artwork,
while still benefiting from the disk cache.

Create addBitmapTo{Memory,Disk}Cache methods, reimplement addBitmapToCache
in terms of them, and only add the processed image to the memory cache.