Call setLoadingImage() once per mImageFetcher, not every getAdapterView().
The previous code loaded and decoded ICON_PENDING_ARTWORK every time
getAdapterView() was called, even if the final image to be displayed
was already in the cache or otherwise available.

This resulted in a ~ 80K unnecessary allocation on every call to
getAdapterView(), which happens ~ 3 times per list row when the list is
scrolled.  Move the list 5 rows and over 1M of memory is allocated,
leading to additional garbage collections and stalls when scrolling
the list.