Slide in activities that list hierarchical collections from the right.
To try and emphasise the spatial dimension in the navigation, animate the
activity transition when viewing music folders, radio stations, and
favourites by sliding the new activity in from the right when it starts,
and back out to the right if the back button is pressed.

Android provides public android.R.anim animations for the inverse effect,
but not for this one although they're in the Android Open Source Project,
so copy the animations from there.

Adjust how item lists are shown.  Remove the "Loading..." string, per the
guides at http://developer.android.com/design/building-blocks/progress.html
and remove the custom background colour, as the transition was jarring.