Use a marquee effect for long strings of track info.
The previous code would wrap long track/album/artist information on to
multiple lines when showing "Now Playing" info.  This could cause the size
of the displayed album artwork to shrink/grow depending on available space,
and was visually jarring.

Force the TextViews to be a single line high, and use the "marquee" effect
to scroll items of text.