Restore functionality to show loading progress in list activities.
As the activity may start with cached state this cannot blindly show the
progress indicator.  Adjust SqueezerItemListActivity::maybeOrderPage() to
return a boolean indicating whether or not the page was ordered.

Override this in SqueezerBaseListActivity::maybeOrderPage(), and if page 0
was requested then hide the listview and make the loading progress indicator
visible.