Fix a bug where searches from the search activity did not work.
The previous code called maybeOrderPage(0), but since page 0 had been fetched
it was not refetched, and so second and subsequent attempts to search would
leave the loading view spinning indefinitely.

Fix this by calling clearAndReOrderItems() instead.  Do this after the
loading label has been made visible so that the clearing of the results
does not happen visibly to the user.

Add some comments while I'm here.