Fix a bug where all option menu items were being shown.
Different versions of Android call onCreateOptionsMenu() at different times,
which may be either before or after the service connection has completed.

The code that toggles state checks to see if menu_item_connect is
non-null, and only toggles visibility if that is the case.  This meant that
the service connection might complete before onCreateOptionsMenu() ran,
menu_item_connect was null, and no toggling would take place.  This resulted
in (for example) the "Connect" menu item being visible even after connection
had completed.

Fix this by moving the toggling code in its own method and call it after
service connection has completed and in onCreateOptionsMenu.  This ensures
that the menu items are set correctly at either point.

While I'm here, rename the variables that reference the items, as the
"Button" suffix was misleading.