Implement album grid layout
Album list layout is now selectable between the existing list view and a new
grid view. The grid view organizes the albums in a grid, and uses larger
artwork.

The selection is done from the options menu. The existing menu entry for
choosing the sort order of the albums, is renamed to "View", and the resulting
dialog is used for both selecting the layout and the sort order. This is
modelled after the view selection on the LMS web interface (and other android
apps). The selected layout is stored in preferences and used in all subsequent
displays of the album list. Like the sort order strings, the strings for
selecting the layout is fetched from the server.

The default layout (i.e. when not previously selected by the user), is based on
the screen size, on the assumption that the artwork grid is preferred on larger
screens.

To implement this the item_list layout is extended with an optional GridView.
To avoid too many setVisibility(GONE/VISIBLE), the optional views (ListView/
GridView/Loading) are put in a ViewAnimator. The BaseListActivity now takes
an extra parameter in its intent, which chooses the current list view.

To support the larger artwork, the image fetcher used in the item lists can now
be overridden. This is accomplished by making the createImageFetcher method
protected, and make it return the newly created image fetcher. This is the same
way the item view can be specified for an item list.

To make it easier to implement the grid item view, the
SqueezerAlbumArtView.getAdapterView method got an overridden method with an
ekstra parameter, layoutId, which specifies the layout to use. Now we just need
to create a new layout resource for the grid items, and specify it in the new
item view logic.

On a side note the image fetcher should be embedded in the item view logic,
since the item view inflates the layout, and therefore already knows about the
image size. Also they are both created from the item list activity. This will
also eliminate the need for all adapters to have an image fetcher.

This new implementation revealed a problem when switching the orientation in the
album list. Using a ViewAnimator caused views to be drawn in onCreate (measure),
which is before the service is connected, so we cant load images.

To fix this the adapter is not set on the listview until after the service is
connected (onServiceConnected/onResume). But that caused the list to scroll to
top, so the current position in the list is retained.

The item view is referenced from the adapter, so it is indirectly retained, so
is removed from the retain fragment.