Properly return the boolean value for context menu clicks from search results.
Without this, Android assumes that the context menu click was not handled,
and it bubbles up to the next context menu handler, which is in the
NowPlayingFragment.  The end result is that you choose a context item like
"Browse songs on album" and it shows the album selected from the search
results, and immediately loads the album that's currently playing on top of
that.