Fixes the activity title when viewing favourites.
The activity title for favourites was previously "0 null" (or some other
number) because SqueezerPluginItemListActivity::createItemView() returns
a SqueezerPluginItemView, and its getQuantityString() method always returns
null.

This adds a specialisation of both of those classes for favourites, to
return a list activity and view class with a custom quantity string.

Adjust SqueezerHomeActivity accordingly.

Create new string resources for this, including a specialisation for the
correct spelling of "Favourites"