Handle events that invalidate the current playlist.
When moving through the playlist the server might add or delete tracks
from it.  For example, if it's a random mix.  If it does this it sends
"addtracks" and "delete" notifications.  Listen for these, and if found,
call a callback in the activity to handle them.

Provide a basic callback for these that invalidates the playlist and
re-orders it from the server, to ensure that the UI is consistent.