Fix item list problems with orientation change
The adapter is retained in onCreate so the list view will start generating the
views. This causes various problems, because the service is not yet connected.

To fix this the adapter is not set on the listview until after the service is
connected (onServiceConnected/onResume). But that caused the list to scroll to
top, so the current position in the list is retained.