Do not call service to get a server string
Instead take advantage that java enums are singletons, and just grab
the server string directly from the enum.
The strings are ordered from the server when a connection is made, and
written to the enum be the service.

This fixes a crash when switching orientation with the album display
options dialog open, because it contains server strings, and the
dialog is displayed before the service is connected.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>