Enable Proguard for gradle builds.
Split the Proguard configuration in to several files, one for
Android defaults ("proguard.cfg"), one per library that has special
Proguard requirements ("proguard-acra.cfg", "proguard-guava.cfg")
and one for things that are specific to Squeezer.

Guava needs the "jsr305" library installed to work with Proguard,
so add that as a dependency.

Customise the "release" buildType to run Proguard and use these
configuration files.

Add a signingConfig to the "live" product flavour so that it shows up
in the list of tasks.  Use the debug signing config for the moment.