Update gradle config to support signing releases.
Create gradle.properties, which defines a property file called
'squeezer.properties'.  This file (not checked in to source control)
contains two entries:

  key.store.password=...
  key.alias.password=...

defining the values for those passwords.

Update build.gradle to look for this file (indirectly, through the
value of "Squeezer.properties") and use the values from it if it exists.
If it doesn't exist then provide dummy values to allow the other tasks
to complete.