Focus scrollable TextViews instead of selecting them.
On (at least) devices running API 7, calling setSelected() on the TextView
displays the text in a darker colour that's difficult to read.  Marking them
as focused does not have this problem.

The trade-off is that only one TextView can have the focus, so this disables
scrolling for TextViews other than the track name.