Add a local build setup
to make "./gradlew build" work on a fresh clone, even if you don't
have the passwords for the official release key.
This will also enable any developer to test the release version of the
application.

When keys are not provided in squeezer.properties the app will be build
just like the official release version, but signed with a local release key
with password "squeezer" for the key and the keystore.