Parse status line from LMS more conservatively
This will fix a reported crash from bugsense.

When the status line come, from some servers  or in some circomstances,
the token for the playing mode (play/pause/stop), may not present.

To handle this and because this is dependant on the server version, and
at any rate beyond our control, Squeezer now checks if mode, shuffle and
repeat status are present, and if not, doesn't update the internal player
state or the UI.