Implement grid layout for song list.
Preferred layout is selectable from view menu. Default is list layout.
The view dialog shares functionality with the album view menu, so common
code is moved to the new BaseViewDialog.

There was no server string for song display options, so added the translation
for album display options where the word for album is substituted with the word
for song.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>