Do not use AIDL for service calls and asynchronous callbacks
As recommended by <http://developer.android.com/guide/components/bound-services.html>
and <http://developer.android.com/guide/components/aidl.html>

This make the code simpler, and the app faster and less memory consuming,
because we do not need to marshall the objects to and from the service.

Note we still use AIDL for RemoteCallbackList (subscriptions), because of the
generic boundary.