Implement player dropdown in the actionbar
In the service callback to observers, when the list of players is learned
from the server, and when changes to the player list is notified. A new
handler id registered to receive "client" notifications.

In the now playing fragment (which is present in all activites), when the
player list is notified, set up a dropdown with the known players, and the
active player selected. When a player is selected from the dropdown, call
the service with the selected player.

Remove all updates of the activity title (since it is overridden by the
action bar). The number of items is no longer showed, and for radios and
musicfolders, update the header of the item list, instead of the activity
title.

Signed-off-by: Kurt Aaholst <kaaholst@gmail.com>