Updated SqueezePlayer integration
Add extras to the intent for the SqueezePlayer service, to ensure that
Squeezer and SqueezePlayer are running against the same server.

Send a keep alive intent every 10 minutes when a Squeezer activity is
active. This is to ensure that the LMS can see the player, and keep it
from removing it from the player-list.

Also made the handling of the server preferences slightly more object
oriented.